Updated code for checkings

+++ earning/EarningService.js
@@ -139,7 +139,7 @@ async function deleteEarningById(req, res) {
     const { earningId } = req.body;
     let earning = await Earning.findById(earningId);
     // console.log(earning._id)
-    if (earning) {
+    if (earning && earning.type == 'other') {
       let isDeleted = await Earning.deleteOne({ _id: earningId });
       res.json({ message: "Deleted earning successfully" });
     } else {
